pyproject.toml
[tool.poetry]
name = "aio-lambda-api"
version = "0.0.0"
description = "Simple AsyncIO AWS lambda HTTP API"
readme = "README.md"
authors = ["JGoutin"]
license = "BSD-2-Clause"
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Typing :: Typed",
]
keywords = ["aws", "lambda", "http", "API", "asyncio"]
repository = "https://github.com/JGoutin/aio_lambda_api"

[tool.poetry.dependencies]
python = "^3.9"
aioboto3 = { version = "*", optional = true }
uvloop = { version = "*", optional = true }
orjson = { version = "*", optional = true }
pydantic = { version = "*", optional = true }
aiohttp = { version = "*", optional = true, extras = ["speedups"] }

[tool.poetry.dev-dependencies]
bandit = "*"
botocore-stubs = "*"
boto3-stubs = {extras = ["ssm"], version = "*"}
black = "*"
flake8 = "*"
flake8-docstrings = "*"
mypy = "*"
pre-commit = "*"
pytest-cov = "*"
radon = "*"
yamllint = "*"
types-urllib3 = "*"
types-setuptools = "*"
types-requests = "*"

[tool.poetry.extras]
all = ["aioboto3", "pydantic", "orjson", "uvloop", "aiohttp"]
aws = ["aioboto3"]
validation = ["pydantic"]
speedups = ["orjson", "uvloop", "aiohttp"]

[tool.pytest.ini_options]
addopts = [
    "--strict-markers",
    "--cov=aio_lambda_api",
]

[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:"
]

[tool.mypy]
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
implicit_reexport = false
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "tests.conftest",
    "aioboto3",
    "uvloop",
]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
